requirements.txt
pandas
openpyxl
supabase
PyJWT
